good text

+++ cub_head.h
@@ -183,7 +183,7 @@ void                no_file(void);
 void                check_map(char **map, int last);
 void                map_err(void);
 void                init_text(t_data *mlx_s);
-void                put_text(int drawStart, int drawEnd, int i, t_data *mlx_s, int lineHeight, int perpWallDists);
+void                put_text(int drawStart, int drawEnd, int i, t_data *mlx_s, double lineHeight, double perpWallDists);
 void                mlx_pixel_get(t_tex *data, int x, int y, unsigned int *color);
 
 #endif
